refactor mysql and sqlite
