+++ db/mysql.py
@@ -1,17 +1,19 @@
 import os
-from .strategy import DatabaseStrategy
+from .database_strategy import DatabaseStrategy
 
 
 class MySQL(DatabaseStrategy):
-    def database_uri(self, mysql_database=None):
-        if mysql_database is None:
-            mysql_database = os.environ.get('MYSQL_DATABASE')
+    mysql_host = os.environ.get('MYSQL_HOST')
+    mysql_user = os.environ.get('MYSQL_USER')
+    mysql_password = os.environ.get('MYSQL_PASSWORD')
+    mysql_database = os.environ.get('MYSQL_DATABASE')
 
-        mysql_host = os.environ.get('MYSQL_HOST')
-        mysql_user = os.environ.get('MYSQL_USER')
-        mysql_password = os.environ.get('MYSQL_PASSWORD')
-        return f'mysql+pymysql://{mysql_user}:{mysql_password}@{mysql_host}/{mysql_database}'  # noqa: E501
+    @property
+    def test_mysql_database(self):
+        return f'{self.mysql_database}_test'
+
+    def database_uri(self):
+        return f'mysql+pymysql://{self.mysql_user}:{self.mysql_password}@{self.mysql_host}/{self.mysql_database}'  # noqa: E501
 
     def test_database_uri(self):
-        mysql_database = f'{os.environ.get('MYSQL_DATABASE')}_test'
-        return self.database_uri(mysql_database)
+        return f'mysql+pymysql://{self.mysql_user}:{self.mysql_password}@{self.mysql_host}/{self.test_mysql_database}'  # noqa: E501
