+++ db/sqlite.py
@@ -1,11 +1,17 @@
 import os
-from .strategy import DatabaseStrategy
+from .database_strategy import DatabaseStrategy
 
 
-class MySQL(DatabaseStrategy):
-    def database_uri(self, database="app.db"):
-        basedir = os.path.abspath(os.path.dirname(__file__))
-        return f'sqlite:///{basedir}/{database}'
+class SQLite(DatabaseStrategy):
+    basedir = os.path.abspath(os.path.dirname(__file__))
+    sqlite_database = "app.db"
+
+    @property
+    def test_sqlite_database(self):
+        return f'{self.sqlite_database}_test'
+
+    def database_uri(self):
+        return f'sqlite:///{self.basedir}/{self.sqlite_database}'
 
     def test_database_uri(self):
-        return self.database_uri('app_test.db')
+        return f'sqlite:///{self.basedir}/{self.test_sqlite_database}'
